requirements.txt
# Core Dependencies
aiXplain>=0.2.30
python-dotenv>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# CLI and UI
click>=8.1.0
streamlit>=1.30.0
rich>=13.0.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
tenacity>=8.2.0

# Logging
loguru>=0.7.0
